Refactor and cleanup pom
