+++ com.intuit.dsl.service.parent/web-client-service-evaluator/pom.xml
@@ -9,7 +9,6 @@
     <version>1.0.0-SNAPSHOT</version>
   </parent>
 
-
   <name>Web Client Service Evaluator</name>
   <groupId>com.intuit.dsl.service.parent</groupId>
   <version>1.0.0-SNAPSHOT</version>
@@ -24,8 +23,8 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <jacoco.ut.execution.data.dir>${project.build.directory}/coverage-reports</jacoco.ut.execution.data.dir>
     <jacoco.ut.execution.data.file>${jacoco.ut.execution.data.dir}/jacoco-ut.exec</jacoco.ut.execution.data.file>
+    <springboot.version>2.4.8</springboot.version>
   </properties>
-
 
   <dependencies>
     <dependency>
@@ -46,7 +45,8 @@
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-webflux</artifactId>
-      <version>2.1.7.RELEASE</version>
+      <version>${springboot.version}</version>
+      <scope>provided</scope>
     </dependency>
     <dependency>
       <groupId>org.apache.commons</groupId>
