+++ com.intuit.dsl.service.parent/web-client-service-evaluator/src/main/java/com/intuit/service/dsl/evaluator/ServiceEvaluator.java
@@ -14,6 +14,7 @@ import com.intuit.dsl.service.Service;
 import com.intuit.service.dsl.JsonUtils;
 import com.intuit.service.dsl.evaluator.exceptions.ServiceDataRetrieverException;
 import com.intuit.service.dsl.evaluator.exceptions.ServiceEvaluatorException;
+import com.intuit.service.dsl.evaluator.expression.ExpressionEvaluator;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.HashMap;
