+++ com.intuit.dsl.service.parent/web-client-service-evaluator/src/main/java/com/intuit/service/dsl/evaluator/utils/ServiceEvaluatorUtils.java
@@ -1,4 +1,4 @@
-package com.intuit.service.dsl.evaluator;
+package com.intuit.service.dsl.evaluator.utils;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
