+++ com.intuit.dsl.service.parent/web-client-service-evaluator/src/main/java/com/intuit/service/dsl/resource/loader/ServiceResourceLoader.java
@@ -1,4 +1,4 @@
-package com.intuit.service.dsl.evaluator;
+package com.intuit.service.dsl.resource.loader;
 
 import org.eclipse.emf.ecore.resource.Resource;
