+++ com.intuit.dsl.service.parent/web-client-service-evaluator/src/test/java/com/intuit/service/dsl/AddBookServicePut.java
@@ -3,6 +3,7 @@ package com.intuit.service.dsl;
 import static java.util.Collections.EMPTY_MAP;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import com.intuit.service.dsl.utils.TestUtil;
 import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
