+++ com.intuit.dsl.service.parent/web-client-service-evaluator/src/test/java/com/intuit/service/dsl/BookByIdService.java
@@ -4,6 +4,7 @@ import static java.util.Collections.EMPTY_MAP;
 
 import com.intuit.dsl.service.Service;
 import com.intuit.service.dsl.evaluator.ServiceEvaluatorRequest;
+import com.intuit.service.dsl.utils.TestUtil;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
