+++ com.intuit.dsl.service.parent/web-client-service-evaluator/src/test/java/com/intuit/service/dsl/BooksService.java
@@ -5,7 +5,7 @@ import static java.util.Collections.EMPTY_MAP;
 import com.google.common.collect.ImmutableMap;
 import com.intuit.dsl.service.Service;
 import com.intuit.service.dsl.evaluator.ServiceEvaluatorRequest;
-import java.util.Collections;
+import com.intuit.service.dsl.utils.TestUtil;
 import java.util.Optional;
 import reactor.util.context.Context;
