--- com.intuit.dsl.service.parent/web-client-service-evaluator/src/test/java/com/intuit/service/dsl/JsonUtilsTest.java
@@ -1,14 +1,0 @@
-package com.intuit.service.dsl;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-
-import org.junit.Test;
-
-public class JsonUtilsTest {
-
-  @Test
-  public void canExtractArgumentsWithVariables() {
-    assertThat(JsonUtils.MAPPER).isNotNull();
-  }
-
-}
