+++ com.intuit.dsl.service.parent/web-client-service-evaluator/src/test/java/com/intuit/service/dsl/evaluator/ExpressionEvaluatorTest.java
@@ -11,7 +11,8 @@ import com.intuit.dsl.service.StringLiteral;
 import com.intuit.dsl.service.Variable;
 import com.intuit.dsl.service.VariableReference;
 import com.intuit.service.dsl.AddBookServicePost;
-import com.intuit.service.dsl.TestUtil;
+import com.intuit.service.dsl.utils.TestUtil;
+import com.intuit.service.dsl.evaluator.expression.ExpressionEvaluator;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.Before;
