+++ com.intuit.dsl.service.parent/web-client-service-evaluator/src/test/java/com/intuit/service/dsl/evaluator/ServiceEvaluatorTest.java
@@ -24,8 +24,8 @@ import com.intuit.service.dsl.AddBookServicePut;
 import com.intuit.service.dsl.BookByIdService;
 import com.intuit.service.dsl.BooksService;
 import com.intuit.service.dsl.JsonUtils;
-import com.intuit.service.dsl.TestUtil;
 import com.intuit.service.dsl.evaluator.exceptions.ServiceEvaluatorException;
+import com.intuit.service.dsl.utils.TestUtil;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
